Remove all code in binary
The `main.rs` file was used to test early on, but now the `examples` are
used for that purpose. They demonstrate how the library can be used by
another crate and are a better placement for what what previously in
`main.rs`.

The idea of the `main.rs` binary version of the crate going forward is
that it will hold a CLI for interacting with a running Parsnip system.
For example to manually send commands or inspect the state of the queue.